+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/SpringContext.java
@@ -12,7 +12,7 @@ public class SpringContext implements ApplicationContextAware {
 
     /**
      * Returns the Spring managed bean instance of the given class type if it exists. Returns null otherwise.
-     * 
+     *
      * @param beanClass
      * @return
      */
@@ -21,9 +21,9 @@ public class SpringContext implements ApplicationContextAware {
     }
 
     @Override
-    public void setApplicationContext(ApplicationContext context) throws BeansException {
+    public void setApplicationContext(ApplicationContext appContext) throws BeansException {
 
         // store ApplicationContext reference to access required beans later on
-        SpringContext.context = context;
+        SpringContext.context = appContext;
     }
 }