+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/certifiedproduct/service/ListingService.java
@@ -103,13 +103,12 @@ public class ListingService {
     public CertifiedProductSearchDetails createCertifiedSearchDetails(Long listingId) throws EntityRetrievalException {
 
         CertifiedProductDetailsDTO dto = certifiedProductSearchResultDAO.getById(listingId);
-        CertifiedProductSearchDetails searchDetails = createCertifiedProductSearchDetailsBasic(certifiedProductSearchResultDAO.getById(listingId));
+        CertifiedProductSearchDetails searchDetails = createCertifiedProductSearchDetailsWithBasicDataOnly(certifiedProductSearchResultDAO.getById(listingId));
 
         searchDetails.setCertificationResults(certificationResultService.getCertificationResults(searchDetails));
         searchDetails.setCqmResults(cqmResultsService.getCqmResultDetails(dto.getId(), dto.getYear()));
         searchDetails.setCertificationEvents(certificationStatusEventsService.getCertificationStatusEvents(dto.getId()));
         searchDetails.setMeaningfulUseUserHistory(meaningfulUseUserHistoryService.getMeaningfulUseUserHistory(dto.getId()));
-        searchDetails = populateDirectReviews(searchDetails);
 
         // get first-level parents and children
         searchDetails.getIcs().setParents(populateRelatedCertifiedProducts(getCertifiedProductParents(dto.getId())));
@@ -119,11 +118,11 @@ public class ListingService {
 
     public CertifiedProductSearchDetails createCertifiedProductSearchDetailsBasic(Long listingId) throws EntityRetrievalException {
         CertifiedProductDetailsDTO dto = certifiedProductSearchResultDAO.getById(listingId);
-        return createCertifiedProductSearchDetailsBasic(dto);
+        return createCertifiedProductSearchDetailsWithBasicDataOnly(dto);
     }
 
-    public CertifiedProductSearchDetails createCertifiedProductSearchDetailsBasic(CertifiedProductDetailsDTO dto) throws EntityRetrievalException {
-        CertifiedProductSearchDetails searchDetails = CertifiedProductSearchDetails.builder()
+    public CertifiedProductSearchDetails createCertifiedProductSearchDetailsWithBasicDataOnly(CertifiedProductDetailsDTO dto) throws EntityRetrievalException {
+        CertifiedProductSearchDetails listing = CertifiedProductSearchDetails.builder()
                 .id(dto.getId())
                 .acbCertificationId(dto.getAcbCertificationId())
                 .certificationDate(dto.getCertificationDate() != null ? dto.getCertificationDate().getTime() : null)
@@ -171,8 +170,10 @@ public class ListingService {
 
         InheritedCertificationStatus ics = new InheritedCertificationStatus();
         ics.setInherits(dto.getIcs());
-        searchDetails.setIcs(ics);
-        return searchDetails;
+        listing.setIcs(ics);
+
+        listing = populateDirectReviews(listing);
+        return listing;
     }
 
     private List<CertifiedProductTestingLab> getTestingLabs(Long listingId) throws EntityRetrievalException {